Using cartopy, performance tweaks, updated notes
This commit addresses a couple of issues. First, cartopy has been used
to plot the data on maps in lieu of basemap. This should be more
relevant to people going forward. I will keep around a branch that
contains the basemap version of the notebook, but not updates will
occur on that branch barring any unforeseen errors.

Secondly, per #4, it was pointed out that passing Fortran memory order
arrays instead of C memory order arrays is faster, see the issue from
JiaweiZhuang/xESMF#6 for further details. Care was taken to make sure
code logic is still intuitive.

Closes #3
Closes #4